requirements.txt
tensorflow
numpy
pandas
scikit-learn
tqdm
matplotlib
Pillow
